complete reorganisation: home is now home
this is not initialised in the home directory (~), and so the files
should be named accordingly. we also need to move a lot of stuff about.
this allows us not to have to symlink everythin (which is annoying) or
source multiple different files (also annoying).

I've added a whole bunch of sudmodules, so I can keep track of the tools
I'm actually using and keep the structure consistent.